+++ backend/src/main/java/com/qualinterns/dealership_app_api/controller/VehicleController.java
@@ -2,6 +2,7 @@ package com.qualinterns.dealership_app_api.controller;
 
 import com.qualinterns.dealership_app_api.dto.VehicleDto;
 import com.qualinterns.dealership_app_api.service.VehicleService;
+import io.github.bucket4j.Bucket;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -24,12 +25,15 @@ import java.util.Map;
 public class VehicleController {
 
     private VehicleService vehicleService;
+    public Bucket bucket;
 
     @GetMapping
-    public ResponseEntity<Map<String, Object>> getAllVehicles(@RequestParam(name = "available", required = false) boolean available,
+    public ResponseEntity<?> getAllVehicles(@RequestParam(name = "available", required = false) boolean available,
                                                               @RequestParam(defaultValue = "0") int page,
                                                               @RequestParam int size) {
         try {
+            if (!bucket.tryConsume(1))
+                return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Rate limit exceeded. Try again later.");
             Pageable pageable = size == 0 ? Pageable.unpaged() : PageRequest.of(page, size);
             Page<?> vehicles;
             Map<String, Object> response = new HashMap<>();
@@ -49,8 +53,10 @@ public class VehicleController {
     }
 
     @GetMapping("/{make}")
-    public ResponseEntity<List<VehicleDto>> getVehicleByMake(@PathVariable String make) {
+    public ResponseEntity<?> getVehicleByMake(@PathVariable String make) {
         try {
+            if (!bucket.tryConsume(1))
+                return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Rate limit exceeded. Try again later.");
             var vehicles = vehicleService.getVehicleByMake(make);
             if (vehicles.isEmpty()) {
                 return ResponseEntity.notFound().build();
@@ -63,8 +69,10 @@ public class VehicleController {
     }
 
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<VehicleDto> createVehicle(@ModelAttribute @Validated VehicleDto vehicleDto) {
+    public ResponseEntity<?> createVehicle(@ModelAttribute @Validated VehicleDto vehicleDto) {
         try {
+            if (!bucket.tryConsume(1))
+                return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Rate limit exceeded. Try again later.");
             vehicleService.createVehicle(vehicleDto);
             return ResponseEntity.status(HttpStatus.CREATED).build();
         } catch (IllegalArgumentException e) {
@@ -80,6 +88,8 @@ public class VehicleController {
     @PatchMapping("/{vehicleId}")
     public ResponseEntity<?> updateVehicle(@PathVariable int vehicleId, @RequestBody @Validated VehicleDto newVehicleDetails) {
         try {
+            if (!bucket.tryConsume(1))
+                return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Rate limit exceeded. Try again later.");
             VehicleDto updatedVehicle = vehicleService.updateVehicle(vehicleId, newVehicleDetails);
             return ResponseEntity.ok(updatedVehicle);
         } catch (EntityNotFoundException e) {
@@ -93,6 +103,8 @@ public class VehicleController {
     @DeleteMapping("/{vehicleId}")
     public ResponseEntity<?> deleteVehicle(@PathVariable int vehicleId) {
         try {
+            if (!bucket.tryConsume(1))
+                return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Rate limit exceeded. Try again later.");
             vehicleService.deleteVehicle(vehicleId);
             return ResponseEntity.noContent().build();
         } catch (DataIntegrityViolationException e) {
